requirements.txt
streamlit
scikit-learn
pandas
numpy
requests
python-dotenv
pickle4

